+++ src/CreateRecordsPlugin.js
@@ -7,6 +7,7 @@ import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
 import { Button } from '@folio/stripes/components';
+import { IfPermission } from '@folio/stripes/core';
 
 import DataProvider from './providers/DataProvider';
 import CreateRecordsWrapper from './CreateRecordsWrapper';
@@ -16,7 +17,7 @@ const CreateRecordsPlugin = ({
   open,
   onOpen,
   onClose,
-  buttonVisible,
+  renderTrigger,
 }) => {
   const [isModalOpen, toggleModal] = useState(false);
 
@@ -38,39 +39,45 @@ const CreateRecordsPlugin = ({
 
   useEffect(() => toggleModal(open), [open]);
 
+  const triggerProps = {
+    'data-test-add-inventory-records': true,
+    id: 'add-inventory-record-trigger-btn',
+    onClick: openModal,
+  };
+
   return (
-    <>
+    <IfPermission perm="ui-plugin-create-inventory-records.create">
       {
-        !isModalOpen && buttonVisible &&
-        <Button
-          data-test-add-inventory-records
-          buttonStyle={buttonStyle}
-          marginBottom0
-          onClick={openModal}
-        >
-          <FormattedMessage id="ui-plugin-create-inventory-records.fastAddLabel" />
-        </Button>
+        !isModalOpen &&
+          (renderTrigger ? renderTrigger(triggerProps) : (
+            <Button
+              buttonStyle={buttonStyle}
+              marginBottom0
+              {...triggerProps}
+            >
+              <FormattedMessage id="ui-plugin-create-inventory-records.fastAddLabel" />
+            </Button>
+          ))
       }
       {isModalOpen &&
         <DataProvider>
           <CreateRecordsWrapper onClose={closeModal} />
         </DataProvider>
       }
-    </>
+    </IfPermission>
   );
 };
 
 CreateRecordsPlugin.defaultProps = {
-  buttonVisible: true,
   open: false,
 };
 
 CreateRecordsPlugin.propTypes = {
   buttonStyle: PropTypes.string,
-  buttonVisible: PropTypes.bool,
   open: PropTypes.bool,
   onOpen: PropTypes.func,
   onClose: PropTypes.func,
+  renderTrigger: PropTypes.func,
 };
 
 export default CreateRecordsPlugin;